Make generic TableField and move styles
This is so we have a base for other table rows that use different interactive elements
